+++ app/src/main/java/ru/firstSet/kotlinOne/Data/MovieApi.kt
@@ -12,10 +12,6 @@ import retrofit2.http.GET
 import retrofit2.http.Path
 import ru.firstSet.kotlinOne.ResultGenre
 
-private const val TAG = "RetrofitMovie"
-private const val apiKey = "f1eaa713b8b88ceef63a9cd8be1f7920"
-private const val BASE_URL = "https://api.themoviedb.org/3/"
-
     interface MoviesApi {
         @GET("genre/movie/list?&language=ru-ru")
         suspend fun getSearchGenre(): ResultGenre
@@ -29,34 +25,3 @@ private const val BASE_URL = "https://api.themoviedb.org/3/"
         @GET("movie/{seachMovie}?language=ru-ru&query=2&include_adult=false")
         suspend fun getMovie(@Path("seachMovie") seachMovie: String): ResultMovie
 }
-
-class MoviesApiHeaderInterceptor : Interceptor {
-    override fun intercept(chain: Interceptor.Chain): Response {
-        val originalRequest = chain.request()
-        val originalHttpUrl = originalRequest.url
-        val request = originalHttpUrl.newBuilder()
-            .addQueryParameter("api_key", apiKey)
-            .build()
-        val url = originalRequest.newBuilder().url(request).build()
-        return chain.proceed(url)
-    }
-}
-
-private val client = OkHttpClient().newBuilder()
-    .addInterceptor(HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY))
-    .addInterceptor(MoviesApiHeaderInterceptor())
-    .build()
-
-private val json = Json {
-    ignoreUnknownKeys = true
-}
-
-@Suppress("EXPERIMENTAL_API_USAGE")
-private val retrofit: Retrofit = Retrofit.Builder()
-    .client(client)
-    .baseUrl(BASE_URL)
-    .addConverterFactory(json.asConverterFactory("application/json".toMediaType()))
-    .build()
-
-public val moviesApi: MoviesApi = retrofit.create(MoviesApi::class.java)
-
